requirements.txt
# Core FastAPI dependencies
fastapi==0.116.1
uvicorn[standard]==0.32.1
pydantic==2.10.3

# AI/ML dependencies
haystack-ai==2.10.0
google-generativeai==0.8.3
tavily-python==0.5.0
sentence-transformers==3.3.1

# Data processing
numpy==2.1.3
pandas==2.2.3

# Template engine
jinja2==3.1.4

# Environment management
python-dotenv==1.0.1

# PII Detection and Security
presidio-analyzer==2.2.359
presidio-anonymizer==2.2.359
spacy==3.8.2
